+++ lib/src/themes/texts/ds_text_theme.theme.dart
@@ -1,3 +1,4 @@
+import 'package:blip_ds/src/themes/colors/ds_colors.theme.dart';
 import 'package:blip_ds/src/themes/texts/styles/ds_body_text_style.theme.dart';
 import 'package:blip_ds/src/themes/texts/styles/ds_button_text_style.theme.dart';
 import 'package:blip_ds/src/themes/texts/styles/ds_caption_small_text_style.theme.dart';
@@ -8,12 +9,15 @@ import 'package:flutter/material.dart';
 
 /// A [TextTheme] used by Material Design to automatically apply our Design System's [TextStyle].
 class DSTextTheme extends TextTheme {
+  /// Creates a Design System's [TextTheme].
   const DSTextTheme()
       : super(
           headline1: const DSHeadlineLargeTextStyle(),
           headline2: const DSHeadlineSmallTextStyle(),
           bodyText1: const DSBodyTextStyle(),
-          button: const DSButtonTextStyle(),
+          button: const DSButtonTextStyle(
+            color: DSColors.neutralDarkCity,
+          ),
           subtitle1: const DSCaptionTextStyle(),
           subtitle2: const DSCaptionSmallTextStyle(),
         );