+++ lib/src/widgets/chat/ds_text_message_bubble.widget.dart
@@ -7,12 +7,14 @@ class DSTextMessageBubble extends StatefulWidget {
   final String text;
   final DSAlign align;
   final List<DSBorderRadius> borderRadius;
+  final bool groupWithPreviousMessage;
 
   const DSTextMessageBubble({
     Key? key,
     required this.text,
     required this.align,
     this.borderRadius = const [DSBorderRadius.all],
+    this.groupWithPreviousMessage = false,
   }) : super(key: key);
 
   @override
@@ -36,14 +38,9 @@ class _DSTextMessageBubbleState extends State<DSTextMessageBubble> {
     return DSMessageBubble(
       align: widget.align,
       borderRadius: widget.borderRadius,
-      child: AnimatedSize(
-        alignment: Alignment.topCenter,
-        curve: Curves.linearToEaseOut,
-        duration: const Duration(milliseconds: 300),
-        clipBehavior: Clip.none,
-        child: Obx(
-          () => _buildText(),
-        ),
+      groupWithPreviousMessage: widget.groupWithPreviousMessage,
+      child: Obx(
+        () => _buildText(),
       ),
     );
   }