+++ sample/lib/main.dart
@@ -1,8 +1,11 @@
 import 'package:blip_ds/blip_ds.dart';
 import 'package:flutter/material.dart';
 import 'package:sample/widgets/showcase/sample_dialog.showcase.dart';
-import 'package:sample/widgets/showcase/sample_message_bubble_showcase.dart';
-import 'package:sample/widgets/showcase/sample_text_style_showcase.dart';
+import 'package:sample/widgets/showcase/sample_button.showcase.dart';
+import 'package:get/route_manager.dart';
+import 'package:sample/widgets/showcase/sample_message_bubble.showcase.dart';
+import 'package:sample/widgets/showcase/sample_text_style.showcase.dart';
+import 'package:sample/widgets/showcase/sample_typing.showcase.dart';
 
 void main() {
   runApp(const SampleApp());
@@ -13,7 +16,7 @@ class SampleApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return MaterialApp(
+    return GetMaterialApp(
       debugShowCheckedModeBanner: false,
       title: 'Blip Design System Showcase',
       theme: ThemeData(
@@ -40,6 +43,10 @@ class HomePage extends StatelessWidget {
             SampleMessageBubbleShowcase(),
             const Divider(color: DSColors.neutralDarkCity),
             const SampleTextStyleShowcase(),
+            const Divider(color: DSColors.neutralDarkCity),
+            const SampleButtonShowcase(),
+            const Divider(color: DSColors.neutralDarkCity),
+            const SampleTypingShowcase(),
             const SampleDialogShowcase(),
           ],
         ),